glue/requirements.txt
boto3==1.38.10
botocore==1.38.10
python-dotenv==1.1.0
PyYAML==6.0.2
psycopg2-binary==2.9.10
requests==2.32.3
Faker==37.1.0
tqdm==4.67.1